Fix plant visual bug
Firefox (for some reason) forced the image to be 0x22px, which meant
that the sprite didn't line up properly. Fixing this by using height and
width rather than the padding workaround meant that Chrome started
showing an image border (because it didn't have a src).

In future I may use `object-fit` and `object-position`, but these are
currently unsupported on IE and Edge. Not that I care all that much
about those browsers...